+++ web/usuaris/src/components/molecules/RoleFilter.tsx
@@ -8,8 +8,9 @@ import {
   type TDropdownOption,
 } from '@itacademy/ui'
 import { useTranslation } from 'react-i18next'
-import { TRole } from '../../types/types'
-import { roles } from '../../constants'
+import { useLocation } from 'react-router-dom'
+import { TRole, UserRole } from '../../types'
+import { paths, roles } from '../../constants'
 
 const StyledDropdown = styled(Dropdown)`
   && button {
@@ -36,6 +37,7 @@ type TRoleFilter = {
 
 export const RoleFilter: FC<TRoleFilter> = ({ handleRole }) => {
   const { t } = useTranslation()
+  const { pathname } = useLocation()
 
   const handleSelectedValue = (selectedOption: TDropdownOption | undefined) => {
     if (selectedOption) {
@@ -49,6 +51,15 @@ export const RoleFilter: FC<TRoleFilter> = ({ handleRole }) => {
       handleRole(undefined)
     }
   }
+
+  if (pathname === paths.mentors)
+    return (
+      <StyledDropdown
+        options={roles}
+        placeholder={t(UserRole.MENTOR)}
+        disabled
+      />
+    )
 
   return roles && roles.length > 0 ? (
     <StyledDropdown