+++ web/usuaris/src/constants/index.ts
@@ -1,4 +1,5 @@
-export { urls } from './urls'
+export { languages, type Language } from './languages'
 export { paths } from './paths'
 export { roles } from './roles'
-export { languages, type Language } from './languages'
+export { sections } from './sections'
+export { urls } from './urls'